More error handling with anyhow and thiserror crates
