+++ Cargo.toml
@@ -35,3 +35,4 @@ hyper-tls = "0.5.0"
 serde_json = "1.0.75"
 derive_more = "0.99.17"
 thiserror = "1.0.30"
+anyhow = "1.0.53"
