+++ src/controllers/health_check.rs
@@ -1,9 +1,9 @@
 use hyper::{Response, StatusCode, Body};
 
-use crate::helpers::response::{ApiResponse, ApiResponseBody};
+use crate::helpers::response::{ApiResponse, ApiResponseBody, ApiBody};
 
 
-pub fn health_check() -> ApiResponse {
+pub fn health_check() -> ApiResponse<ApiBody> {
     let ok_body = Body::from(ApiResponseBody::new("Ok".to_string(), Some("".to_string())));
 
     let response_body = Response::builder()
