+++ src/errors.rs
@@ -1,4 +1,2 @@
 pub mod response;
 pub mod twitter_errors;
-pub mod redis_errors;
-pub mod common;
