+++ src/routes/server.rs
@@ -2,6 +2,7 @@ use hyper::{Request, Body, Method};
 use redis::{Client as RedisClient};
 
 use crate::helpers::request::HyperClient;
+use crate::helpers::response::ApiBody;
 // use crate::app::client::AppClient;
 use crate::{helpers::response::ApiResponse};
 use crate::controllers::{not_found, authorize_bot, health_check, handle_redirect};
@@ -12,7 +13,7 @@ pub async fn routes(
     req: Request<Body>, 
     client: HyperClient,
     conn: RedisClient
-) -> ApiResponse {
+) -> ApiResponse<ApiBody> {
     match (req.method(), req.uri().path()) {
         (&Method::GET, "/") => health_check(),
         (&Method::GET, "/enable") => authorize_bot(&client, conn).await,
